Move slashings mock to mock pkg
