+++ beacon-chain/operations/slashings/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "doc.go",
         "log.go",
         "metrics.go",
-        "mock.go",
         "service.go",
         "types.go",
     ],
@@ -44,6 +43,7 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
+        "//beacon-chain/operations/slashings/mock:go_default_library",
         "//beacon-chain/state:go_default_library",
         "//config/fieldparams:go_default_library",
         "//config/params:go_default_library",
