+++ beacon-chain/operations/slashings/mock/BUILD.bazel
@@ -4,11 +4,10 @@ go_library(
     name = "go_default_library",
     testonly = True,
     srcs = ["mock.go"],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/operations/voluntaryexits/mock",
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/operations/slashings/mock",
     visibility = ["//visibility:public"],
     deps = [
         "//beacon-chain/state:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
-        "@com_github_prysmaticlabs_eth2_types//:go_default_library",
     ],
 )
