+++ beacon-chain/operations/slashings/mock/mock.go
@@ -1,4 +1,4 @@
-package slashings
+package mock
 
 import (
 	"context"
