+++ beacon-chain/operations/slashings/service_test.go
@@ -3,14 +3,15 @@ package slashings
 import (
 	"testing"
 
+	"github.com/prysmaticlabs/prysm/beacon-chain/operations/slashings/mock"
 	"github.com/prysmaticlabs/prysm/testing/require"
 )
 
 var (
 	_ = PoolManager(&Pool{})
 	_ = PoolInserter(&Pool{})
-	_ = PoolManager(&PoolMock{})
-	_ = PoolInserter(&PoolMock{})
+	_ = PoolManager(&mock.PoolMock{})
+	_ = PoolInserter(&mock.PoolMock{})
 )
 
 func TestPool_validatorSlashingPreconditionCheck_requiresLock(t *testing.T) {
