+++ testing/endtoend/slasher_simulator_e2e_test.go
@@ -9,7 +9,7 @@ import (
 	types "github.com/prysmaticlabs/eth2-types"
 	mock "github.com/prysmaticlabs/prysm/beacon-chain/blockchain/testing"
 	dbtest "github.com/prysmaticlabs/prysm/beacon-chain/db/testing"
-	"github.com/prysmaticlabs/prysm/beacon-chain/operations/slashings"
+	mockslashings "github.com/prysmaticlabs/prysm/beacon-chain/operations/slashings/mock"
 	mockstategen "github.com/prysmaticlabs/prysm/beacon-chain/state/stategen/mock"
 	"github.com/prysmaticlabs/prysm/crypto/bls"
 	ethpb "github.com/prysmaticlabs/prysm/proto/prysm/v1alpha1"
@@ -91,7 +91,7 @@ func TestEndToEnd_SlasherSimulator(t *testing.T) {
 		AttestationStateFetcher:     mockChain,
 		StateGen:                    gen,
 		PrivateKeysByValidatorIndex: privKeys,
-		SlashingsPool:               &slashings.PoolMock{},
+		SlashingsPool:               &mockslashings.PoolMock{},
 		SyncChecker:                 mockSyncChecker{},
 	})
 	require.NoError(t, err)
